Update instructions on initial.js to make intent clearer
For a script that takes a three part name and generates initials, I would expect the ouput in this case to be "CKJ", but with the instructions as written, I would write a script that would ouput "CREKATJOH". There's some ambiguity here - I've updated the instructions to make a standard 3 char initial variable, but if the intent is to create a 9 char initial set, I would suggest either expanding the instructions further, or moving away from using initials as the example.